check in prior to going back to starting array at 1
muav, musv and gv are 2d arrays and I added the dim of wl. At first I matched it to the convention of the other dim, namely starting at 1, but that got messed up and starting the wl dim at 0 seemed smoother. However, since there is a follow-on Matlab step writing out the props, this must be changed back